package.py - Connect to Terraform's remapped TTY if it's available
It seems that Terraform remaps all TTY FDs to a range {3..5}
and this internal feature used by terrafrom plugins for I/O.

So the package.py just tries to reuse Terraform's TTY if it's available.
This feature reused by package.py to make logging output from the
prepare stage.

On the prepare stage all logging happens only on a debug level, so to
see something use Terraform with set env variable:
  > export TF_LAMBDA_PACKAGE_LOG_LEVEL=DEBUG
  > terraform apply